Restaurants

+++ models/restaurantModel.go
@@ -16,7 +16,7 @@ type Restaurant struct {
 	Delivery      bool               `json:"delivery" binding:"required" bson:"delivery"`
 	CreatedAt     time.Time          `json:"created_at,omitempty" bson:"created_at,omitempty"`
 	UpdatedAt     time.Time          `json:"updated_at,omitempty" bson:"updated_at,omitempty"`
-	Rating        int                `json:"rating" binding:"required" bson:"rating"`
+	Rating        float64            `json:"rating" binding:"required" bson:"rating"`
 	RatingCount   int                `json:"ratingCount" binding:"required" bson:"ratingCount"`
 	Menu          []Menu             `json:"menu" binding:"required" bson:"menu"`
 }
